+++ tests/testthat/test-generate_rules_matrix.R
@@ -8,9 +8,8 @@ test_that("Rules Extracted Correctly", {
   z <- dataset_cont[["z"]]
   X <- dataset_cont[["X"]]
   ite_method <- "bart"
-  include_ps <- "TRUE"
-  ps_method <- "SL.xgboost"
-  oreg_method <- NA
+  learner_ps <- "SL.xgboost"
+  learner_y <- NA
   ntrees_rf <- 100
   ntrees_gbm <- 50
   node_size <- 20
@@ -28,9 +27,8 @@ test_that("Rules Extracted Correctly", {
 
   # Step 2: Estimate ITE
   ite <- estimate_ite(y, z, X, ite_method,
-                           include_ps = include_ps,
-                           ps_method = ps_method,
-                           oreg_method = oreg_method)
+                           learner_ps = learner_ps,
+                           learner_y = learner_y)
 
   # Step 3: Generate rules list
   initial_rules <- generate_rules(X, ite, ntrees_rf, ntrees_gbm, node_size,